Fix review issues and increase network values
