--- project_1/data/1_sec_queue_size.txt
@@ -1,26 +1,0 @@
-1: 0
-2: 0
-3: 0
-4: 0
-5: 0
-6: 0
-7: 0
-8: 0
-9: 0
-10: 0
-11: 0
-12: 0
-13: 0
-14: 0
-15: 0
-16: 47104
-17: 98304
-18: 98304
-19: 149504
-20: 238583
-21: 334062
-22: 435941
-23: 538341
-24: 640741
-25: 743141
-26: 845541
