+++ project_1/data/30_sec_queue_size.txt
@@ -1,0 +1,27 @@
+0 0 1
+1 0 2
+2 0 4
+3 0 8
+4 0 16
+5 0 32
+6 0 64
+7 0 128
+8 0 256
+9 0 512
+10 0 1024
+11 0 2048
+12 0 4096
+13 0 8192
+14 0 16384
+15 0 32768
+16 0 65536
+17 0 131072
+18 0 262144
+19 0 524288
+20 0 1048576
+21 187940 2097152
+22 6479396 4131657
+23 12729396 5453293
+24 12875880 5453293
+25 19272364 5502121
+26 30269965 5700372
