+++ project_1/src/main.cpp
@@ -3,7 +3,7 @@
 #include <iostream>
 
 int main(int argc, char* argv[]) {
-    std::uint32_t simulation_time_sec = 1;
+    std::uint32_t simulation_time_sec = 30;
     std::vector<std::uint32_t> distances_to_receiver_us = {10, 4, 20};
 
     NetworkSimulator simulator(distances_to_receiver_us);
